+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "cross-env": "7.0.3",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "joi": "^17.9.1",
     "morgan": "^1.10.0",
     "nanoid": "^3.1.15"
